pyproject.toml
[tool.poetry]
name = "rain"
version = "0.1.0"
description = ""
authors = ["Artemiy <renevartemiy@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.4"}
pydantic = {extras = ["email"], version = "^2.9.2"}
alembic = "^1.14.0"
ruff = "^0.7.2"
mypy = "^1.13.0"
dynaconf = "^3.2.6"
asyncpg = "^0.30.0"
uvicorn = "^0.32.0"
typer = "^0.13.0"
sqlalchemy-utils = "^0.41.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.2.0"
websockets = "^14.0"
sqladmin = "^0.20.1"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
